로컬스토리지 비우기

See merge request contentbankfe/contentbank-user!123

+++ client/src/components/contents/createworksheet/Step2.tsx
@@ -44,9 +44,19 @@ export function Step2() {
       console.log('데이터 조회', parsedData);
       setSendLocalData(parsedData);
       // 로컬 스토리지 값 다 받은 뒤 초기화
-      window.opener.localStorage.clear();
+      //window.opener.localStorage.clear();
+      // if (sendLocalData) {
+      //   window.opener.localStorage.clear();
+      // }
     }
   }, []);
+
+  useEffect(() => {
+    if (sendLocalData) {
+      window.opener.localStorage.clear();
+    }
+  }, [sendLocalData]);
+
   console.log(sendLocalData);
   console.log(sendLocalData?.data);
